Fixed compiler errors with MSVC.

+++ core/jbw/gibbs_field.h
@@ -238,7 +238,7 @@ public:
 	{
 #if SAMPLING_METHOD == MH_SAMPLING
 		LOG_ITEM_TYPE_COUNT = (float) log(cache.item_type_count);
-		LOG_N_SQUARED = (float) 2 * log(n);
+		LOG_N_SQUARED = (float) (2 * log(n));
 #endif
 	}
 
@@ -444,8 +444,8 @@ public:
 					log_acceptance_probability += cache.intensity(new_position, item_type);
 
 					/* add log probability of inverse proposal */
-					logarithm.ensure_size(current->items.length + 2);
-					log_acceptance_probability += (float) -logarithm.get(current->items.length + 1);
+					logarithm.ensure_size((unsigned int) current->items.length + 2);
+					log_acceptance_probability += (float) -logarithm.get((unsigned int) current->items.length + 1);
 
 					/* subtract log probability of forward proposal */
 					log_acceptance_probability -= -LOG_ITEM_TYPE_COUNT - LOG_N_SQUARED;
@@ -498,8 +498,8 @@ public:
 				log_acceptance_probability += -LOG_ITEM_TYPE_COUNT - LOG_N_SQUARED;
 
 				/* subtract log probability of forward proposal */
-					logarithm.ensure_size(current->items.length + 1);
-				log_acceptance_probability -= (float) -logarithm.get(current->items.length);
+				logarithm.ensure_size((unsigned int) current->items.length + 1);
+				log_acceptance_probability -= (float) -logarithm.get((unsigned int) current->items.length);
 
 				/* accept or reject the proposal depending on the computed probability */
 				float random = (float) rng() / rng.max();
